refactor: Moved the activity stream custom classes to better package structure
