+++ EquiTrack/EquiTrack/settings/base.py
@@ -442,7 +442,7 @@ ACTSTREAM_SETTINGS = {
     'FETCH_RELATIONS': True,
     'GFK_FETCH_DEPTH': 1,
     'USE_JSONFIELD': True,
-    'MANAGER': 'EquiTrack.stream.CustomDataActionManager',
+    'MANAGER': 'EquiTrack.stream_feed.managers.CustomDataActionManager',
 }
 
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
